rename: 관련 페이지별로 폴더 정리 및 이동
